Upgrade Scintilla from v4.2.0 to v4.4.6
Close #8900, close #9550

+++ scintilla/lexlib/PropSetSimple.h
@@ -19,7 +19,7 @@ public:
 	void Set(const char *key, const char *val, size_t lenKey, size_t lenVal);
 	void SetMultiple(const char *);
 	const char *Get(const char *key) const;
-	int GetExpanded(const char *key, char *result) const;
+	size_t GetExpanded(const char *key, char *result) const;
 	int GetInt(const char *key, int defaultValue=0) const;
 };
